R6 Class for DuckDB Database Interface

An R6 class that provides an interface to a DuckDB database for the evoland package. This class handles database initialization, data commits, and data fetching operations.

Public fields

connection

DBI connection object to the DuckDB database

path

Character string path to the database file

write_mode

Logical indicating if database is opened in write mode Initialize a new evoland_db object

Active bindings

config

Retrieve evoland_config; may be assigned a config (read_evoland_config()) only if no other config is present

coords_t

A coords_t instance; see create_coords_t() for the type of object to assign. Assigning is an upsert operation.

periods_t

A periods_t instance; see create_periods_t() for the type of object to assign. Assigning is an upsert operation.

lulc_meta_t

A lulc_meta_t instance; see create_lulc_meta_t() for the type of object to assign. Assigning is an upsert operation.

lulc_data_t

A lulc_data_t instance; see as_lulc_data_t() for the type of object to assign. Assigning is an upsert operation.

pred_meta_t

A pred_meta_t instance; see create_pred_meta_t() for the type of object to assign. Assigning is an upsert operation.

pred_sources_v

Retrieve a table of distinct predictor urls and their md5sum

pred_data_t_float

A pred_data_t_float instance; see create_pred_data_t() for the type of object to assign. Assigning is an upsert operation.

pred_data_t_int

A pred_data_t_int instance; see create_pred_data_t() for the type of object to assign. Assigning is an upsert operation.

pred_data_t_bool

A pred_data_t_bool instance; see create_pred_data_t() for the type of object to assign. Assigning is an upsert operation.

trans_meta_t

A trans_meta_t instance; see create_trans_meta_t() for the type of object to assign. Assigning is an upsert operation.

trans_preds_t

A trans_preds_t instance; see create_trans_preds_t() for the type of object to assign. Assigning is an upsert operation.

intrv_meta_t

A intrv_meta_t instance; see create_intrv_meta_t() for the type of object to assign. Assigning is an upsert operation.

intrv_masks_t

A intrv_masks_t instance; see as_intrv_masks_t() for the type of object to assign. Assigning is an upsert operation.

trans_models_t

A trans_models_t instance; see create_trans_models_t() for the type of object to assign. Assigning is an upsert operation.

alloc_params_t

A alloc_params_t instance; see as_alloc_params_t() for the type of object to assign. Assigning is an upsert operation.

Methods

Public methods

evoland_db$new()

evoland_db$commit()

evoland_db$fetch()

evoland_db$list_tables()

evoland_db$execute()

evoland_db$row_count()

evoland_db$delete_from()

evoland_db$copy_db()

evoland_db$clone()

Method new()

Usage

evoland_db$new(path, write = TRUE)

Arguments

path

Character string. Path to the DuckDB database file. May also be ":memory:".

write

Logical. Whether to open the database in write mode. Default is TRUE. If FALSE, the database file must already exist.

Returns

A new evoland_db object Commit data to the database

Method commit()

Usage

evoland_db$commit(x, table_name, mode = "upsert")

Arguments

x

Data object to commit.

table_name

Table to target

mode

Character string. One of "upsert" (default), "append", or "overwrite".

Returns

NULL (called for side effects) Fetch data from the database

Method fetch()

Usage

evoland_db$fetch(table_name, where = NULL, limit = NULL)

Arguments

table_name

Character string. Name of the database table or view to query.

where

Character string. Optional WHERE clause for the SQL query.

limit

integerish, limit the amount of rows to return

Returns

A table of the original class List all tables in the database

Method list_tables()

Usage

evoland_db$list_tables()

Returns

Character vector of table names Execute statement

Method execute()

Usage

evoland_db$execute(statement)

Arguments

statement

A SQL statement

Returns

No. of rows affected by statement Get table row count

Method row_count()

Usage

evoland_db$row_count(table_name)

Arguments

table_name

Character string. Name of the database table or view to query.

Returns

No. of rows affected by statement Empty a table

Method delete_from()

Usage

evoland_db$delete_from(table_name, where = NULL)

Arguments

table_name

Character string. Name of the database table or view to delete.

where

Character string, defaults to NULL: delete everything in table.

Returns

No. of rows affected by statement Copy full DB to a different one

Method copy_db()

Usage

evoland_db$copy_db(target_path, source_db)

Arguments

target_path

Character string. Name of the database file to copy to

source_db

Character string. Name of the database to copy from. Defaults to this object's DB path

Method clone()

The objects of this class are cloneable with this method.

Usage

evoland_db$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.